server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
sqlalchemy==2.0.25
pandas==2.1.4
python-dotenv==1.0.0
numpy==1.26.3